% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{acidity}
\alias{acidity}
\title{Acidity dataset
Acidity series}
\format{
...
}
\usage{
acidity
}
\description{
Acidity dataset
Acidity series
}
\references{
\insertRef{Richardson_Green_1997_RJMCMC}{BayesMultiMode}
}
\keyword{datasets}
