% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayes_estimation.R
\name{bayes_estimation}
\alias{bayes_estimation}
\title{Bayesian estimation of mixture distributions}
\usage{
bayes_estimation(
  data,
  K,
  dist,
  a0 = 10,
  A0 = 10 * K,
  b0 = mean(data),
  B0 = (max(data) - min(data))^2,
  c0 = 2.5,
  e0 = 0,
  g0 = 0.5,
  G0 = 100 * 2.5/0.5/(max(data) - min(data))^2,
  h0 = 0,
  H0 = 10,
  n0 = 2,
  N0 = 0.1,
  nb_iter = 2000,
  burnin = nb_iter/2,
  chains = 4,
  cores = 4,
  refresh = 1000,
  ...
)
}
\arguments{
\item{data}{Numeric vector of input values.}

\item{K}{Integer indicating the maximum number of mixture components.}

\item{dist}{String indicating the distribution of the mixture components.
Currently supports "normal", "student" and "skew_normal".}

\item{a0}{(numeric scalar) Mean of the gamma hyperprior used the dirichlet prior. Default is 10.}

\item{A0}{(numeric scalar) Variance of the gamma hyperprior used the dirichlet prior. Default is 10*K.}

\item{b0}{(numeric scalar) Mean of the mean priors. Default is mean(data).}

\item{B0}{(numeric scalar) Numeric value for the variance of the mean priors. Default is R^2 where R = (max(data) - min(data)).}

\item{c0}{(numeric scalar) Numeric value for variance prior. Default is 2.5.}

\item{e0}{(numeric positive scalar) Dirichlet prior parameter. If 0 so an hyperprior is used instead. Default is 0.}

\item{g0}{(numeric scalar) Numeric value for variance inverse gamma hyperprior. Default is 0.5.}

\item{G0}{(numeric scalar) Numeric value for variance inverse gamma hyperprior. Default is 100*2.5/0.5/R^2.}

\item{h0}{(numeric scalar) Mean of the skew parameters priors. Default is 0.}

\item{H0}{(numeric scalar) Variance of the skew parameters priors. Default is 10.}

\item{n0}{(numeric scalar) Mean of the degree of freedom gamma priors. Default is 2.}

\item{N0}{(numeric scalar) Variance of the degree of freedom gamma priors. Default is 2.}

\item{nb_iter}{Number of MCMC iterations. Default is 2000.}

\item{burnin}{...}

\item{chains}{...}

\item{cores}{...}

\item{refresh}{...}

\item{...}{Other arguments passed to `rstan::sampling`.}
}
\value{
An object of class `BayesMixture`.
}
\description{
Bayesian estimation of mixture distributions
}
