% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayes_info.R
\name{bayes_info}
\alias{bayes_info}
\title{Compute the Bayesian Information Criterion}
\usage{
bayes_info(BayesMix)
}
\arguments{
\item{BayesMix}{An object of class \code{bayes_mixture}.}
}
\value{
Returns a list with:
\item{BIC}{Vector of BIC values for each MCMC draw.}
\item{mean_BIC}{Mean BIC}
\item{sd_BIC}{Standard Deviation of the BIC}
}
\description{
Returns the BIC for each draw as well as its posterior mean and standard deviation.
}
\examples{
# Example with galaxy data ================================================
set.seed(123)

# retrieve galaxy data
y <- galaxy

# estimation
bayesmix <- bayes_fit(
  data = y,
  K = 5, # not many to run the example rapidly
  dist = "normal",
  nb_iter = 500, # not many to run the example rapidly
  burnin = 100
)

# BIC
bayes_info(bayesmix)

}
