% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ct47}
\alias{ct47}
\title{X chromosomal macrosatellite repeats ct47}
\format{
A vector of counts with 410 elements.
}
\usage{
ct47
}
\description{
Repeat units that encode for a cancer testis antigen.\cr
Locus (hg18): Xq24 \cr
Unit (kb): 4.8 \cr
Restriction enzyme: EcoRI \cr
Encoded product : cancer testis antigen 47
}
\references{
\insertRef{schaap_genome-wide_2013}{BayesMultiMode}
}
\keyword{datasets}
