% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{d4z4}
\alias{d4z4}
\title{Autosomal macrosatellite repeats d4z4}
\format{
A vector of counts with 410 elements.
}
\usage{
d4z4
}
\description{
Macrosatellite repeats D4Z4 in the subtelomere of chromosome 4q.\cr
Locus (hg18): 4q35.2 \cr
Unit (kb): 3.3 \cr
Restriction enzyme: EcoRI + HindIII/EcoRI + BlnI/XapI \cr
Encoded product : DUX4
}
\references{
\insertRef{schaap_genome-wide_2013}{BayesMultiMode}
}
\keyword{datasets}
