% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discrete_MF.R
\name{discrete_MF}
\alias{discrete_MF}
\title{Mode-finding for mmixture of discrete distributions}
\usage{
discrete_MF(
  mcmc,
  data,
  pars_names,
  dist,
  type = "all",
  pdf_func = NULL,
  show_plot = FALSE
)
}
\arguments{
\item{mcmc}{a vector of estimated mixture parameters.}

\item{data}{Numeric vector of observations.}

\item{pars_names}{Names of the variables mcmc draws variables}

\item{dist}{String indicating the distribution of the mixture components.
Currently supports "normal", "student" and "skew_normal".}

\item{type}{Type of modes, either unique or all (the latter includes flat modes).}

\item{pdf_func}{Pdf or pmf of the mixture components associated with the mcmc draws (if estimation not in-house)}

\item{show_plot}{Show the data and estimated modes.}
}
\value{
A vector estimated modes.
}
\description{
Mode-finding for mmixture of discrete distributions
}
