% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{enzyme}
\alias{enzyme}
\title{Enzyme dataset
Enzyme series}
\format{
...
}
\usage{
enzyme
}
\description{
Enzyme dataset
Enzyme series
}
\references{
\insertRef{Richardson_Green_1997_RJMCMC}{BayesMultiMode}
}
\keyword{datasets}
