% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixed_point.R
\name{fixed_point}
\alias{fixed_point}
\title{Fixed-point algorithm for finding the modes of a gaussian mixture}
\usage{
fixed_point(params, tol_p = 1e-04, tol_x, show_plot = FALSE)
}
\description{
Fixed-point algorithm to find the modes of a gaussian
mixture from Carreira-Perpinan (2000), section 4 equation (10)
https://doi.org/10.1109/34.888716
}
