% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{galaxy}
\alias{galaxy}
\title{Galaxy series}
\format{
An object of class \code{numeric} of length 82.
}
\source{
https://people.maths.bris.ac.uk/~mapjg/mixdata
}
\usage{
galaxy
}
\description{
Velocity at which 82 galaxies in the Corona Borealis region are moving away from our galaxy, scaled by 1000.
}
\references{
\insertRef{Richardson_Green_1997_RJMCMC}{BayesMultiMode}
}
\keyword{datasets}
