% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gibbs_sfm_algos.R
\name{gibbs_SFM}
\alias{gibbs_SFM}
\title{SFM MCMC algorithms to estimate mixtures.}
\usage{
gibbs_SFM(y, K, nb_iter, priors = list(), print = TRUE, dist)
}
\description{
SFM MCMC algorithms to estimate mixtures.
}
\keyword{internal}
