% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gibbs_SFM_normal.R
\name{gibbs_SFM_normal}
\alias{gibbs_SFM_normal}
\title{Bayesian estimation of mixture of normals}
\usage{
gibbs_SFM_normal(
  y,
  K,
  nb_iter,
  a0 = 1,
  A0 = 1/200,
  b0 = median(y),
  B0 = (max(y) - min(y))^2,
  c0 = 2.5,
  e0 = 0.01,
  g0 = 0.5,
  G0 = 100 * g0/c0/B0,
  prt = TRUE
)
}
\arguments{
\item{y}{(a vector of integers) Observations used to fit the model.}

\item{K}{(an integer) Maximum number of mixture components.}

\item{nb_iter}{(an integer) Number of MCMC iterations.}

\item{prt}{print intermediate of the MCMC estimation ? default = TRUE.}
}
\value{
A matrix of MCMC samples
}
\description{
Gibbs sampler for estimating a SFM of normal distributions from xxx 20.. (univariate version)
}
\references{
\insertRef{malsiner-walli_model-based_2016}{BayesMultiMode}\cr
}
\keyword{internal}
