% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gibbs_SFM_poisson.R
\name{gibbs_SFM_poisson}
\alias{gibbs_SFM_poisson}
\title{Estimation of a mixture of Poisson distributions.}
\usage{
gibbs_SFM_poisson(y, K, nb_iter, priors = list(), printing = TRUE)
}
\arguments{
\item{y}{(a vector of integers) Observations used to fit the model.}

\item{K}{(an integer) Maximum number of mixture components.}

\item{nb_iter}{(an integer) Number of MCMC iterations.}

\item{prt}{print intermediate of the MCMC estimation ? default = TRUE.}
}
\value{
mcmc_draws : a (nb_iter x 2K + 1) matrix. Parameter draws from the posterior distribution at each MCMC iteration.
}
\description{
Bayesian estimation of a mixture of Poisson distributions using a Sparse Finite Mixture MCMC algorithm.
}
\references{
\insertRef{malsiner-walli_model-based_2016}{BayesMultiMode}\cr
\insertRef{basturk_bayes_2021}{BayesMultiMode}
}
\keyword{internal}
