% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gibbs_SFM_skew_normal.R
\name{gibbs_SFM_skew_n}
\alias{gibbs_SFM_skew_n}
\title{Bayesian estimation of mixture of skew normal}
\usage{
gibbs_SFM_skew_n(
  y,
  K,
  nb_iter,
  a0 = 1,
  A0 = 1/200,
  b0 = median(y),
  c0 = 2.5,
  C0 = 0.5 * var(y),
  e0 = a0 * A0,
  g0 = 0.5,
  G0 = g0/(0.5 * var(y)),
  D_xipsi = 10,
  prt = TRUE
)
}
\arguments{
\item{y}{(a vector of integers) Observations used to fit the model.}

\item{K}{(an integer) Maximum number of mixture components.}

\item{nb_iter}{(an integer) Number of MCMC iterations.}

\item{prt}{print intermediate of the MCMC estimation ? default = TRUE.}
}
\value{
A matrix of MCMC samples
}
\description{
Gibbs sampler for estimating a SFM of skew normal distributions from xxx 20..
}
\references{
\insertRef{malsiner-walli_model-based_2016}{BayesMultiMode}\cr
}
\keyword{internal}
