% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MEM.R
\name{MEM}
\alias{MEM}
\title{Mode-finding EM algorithm (MEM)}
\usage{
MEM(mcmc, dist, tol_p = 0.001, tol_x, show_plot = FALSE)
}
\description{
From Li, Jia, Surajit Ray, and Bruce G. Lindsay.
"A nonparametric statistical approach to clustering via mode identification."
Journal of Machine Learning Research 8 (2007): 1687-1723.
}
