% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{msr5p}
\alias{msr5p}
\title{DNA Macrosatellite repeat msr5p (4q).
Locus (hg18): 4q35.2; Unit (kb): 3.3; Encoded product: DUX4.}
\format{
A vector of counts with 410 elements.
}
\usage{
msr5p
}
\description{
DNA Macrosatellite repeat msr5p (4q).
Locus (hg18): 4q35.2; Unit (kb): 3.3; Encoded product: DUX4.
}
\references{
\insertRef{schaap_genome-wide_2013}{BayesMultiMode}
}
\keyword{datasets}
