% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_BayesMixture.R
\name{new_BayesMixture}
\alias{new_BayesMixture}
\title{Creating a S3 object of class `BayesMixture`.
This function is helpful for users who want to explore modes in MCMC draws which have not been
derived using the function `bayes_estimation()`.}
\usage{
new_BayesMixture(
  mcmc,
  data,
  K,
  burnin,
  dist = "NA",
  pars_names,
  pdf_func = NULL,
  dist_type
)
}
\arguments{
\item{mcmc}{A matrix of MCMC draws.}

\item{data}{A vector containing the data used for estimating the model and generating the MCMC draws.}

\item{K}{Number of mixture components.}

\item{burnin}{Number of draws to discard as burnin.}

\item{dist}{Distribution family of the mixture components supported by
the package (e.g. "normal", "student", "skew_normal", "shifted_poisson").}

\item{pars_names}{Mapping between the distribution parameters names.
This input is used only if dist_name is invalid or NULL.}

\item{pdf_func}{Pdf or pmf of the mixture components.
This input is used only if dist_name is invalid or NULL.}

\item{dist_type}{Either "continous" or "discrete"}
}
\value{
An object of class `BayesMixture`.
}
\description{
Creating a S3 object of class `BayesMixture`.
This function is helpful for users who want to explore modes in MCMC draws which have not been
derived using the function `bayes_estimation()`.
}
