% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_Mixture.R
\name{new_Mixture}
\alias{new_Mixture}
\title{Creating a S3 object of class \code{Mixture}}
\usage{
new_Mixture(pars, dist = "NA", pdf_func = NULL, dist_type = "NA", data = NULL)
}
\arguments{
\item{pars}{Named vector of mixture parameters.}

\item{dist}{String indicating the distribution of the mixture components; default is "NA".
Currently supports "normal" and "skew_normal"; not needed if pdf_func is provided.}

\item{pdf_func}{Pdf of the mixture components; default is null.}

\item{dist_type}{Either "continuous" or "discrete".}
}
\value{
A list of class \code{Mixture} containing:
\itemize{
 \item{pars}{ - Same as argument.}
 \item{dist}{ - Same as argument.}
 \item{pdf_func}{ - Same as argument.}
 \item{dist_type}{ - Same as argument.}
}
}
\description{
Function for creating an object of class \code{Mixture} which can subsequently be used as argument in [MEM()] and [fixed_point()].
}
\examples{

# Example with the skew normal =============================================
xi = c(0,6)
omega = c(1,2)
alpha = c(0,0)
p = c(0.8,0.2)
params = c(eta = p, xi = xi, omega = omega, alpha = alpha)
dist = "skew_normal"

mix = new_Mixture(params, dist = dist)

# Example with an arbitrary distribution ===================================

pdf_func <- function(x, pars) {
  sn::dst(x, pars["mu"], pars["sigma"], pars["xi"], pars["nu"])
}


mix = new_Mixture(params, pdf_func = pdf_func)

}
