% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_methods.R
\name{plot.bayes_mixture}
\alias{plot.bayes_mixture}
\title{Plot method for \code{bayes_mixture} objects}
\usage{
\method{plot}{bayes_mixture}(x, draws = 250, draw = NULL, bins = 30, alpha = 0.1, ...)
}
\arguments{
\item{x}{An object of class \code{bayes_mixture}.}

\item{draws}{The number of MCMC draws to plot.}

\item{draw}{Plot estimated mixture in draw \code{draw}; note that \code{draws} is discarded. Default is \code{NULL}.}

\item{bins}{(for continuous mixtures) Number of bins for the histogram of
the data. Passed to \code{geom_histogram()}.}

\item{alpha}{transparency of the density lines. Default is 0.1. Should be greater than 0 and below or equal to 1.}

\item{...}{Not used.}
}
\description{
Plot an estimated mixture for a given number of draws with a frequency distribution of the data.
}
