% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_methods.R
\name{plot.BayesMode}
\alias{plot.BayesMode}
\title{Plot modes}
\usage{
\method{plot}{BayesMode}(x, graphs = c("p1", "number", "loc"), ...)
}
\arguments{
\item{x}{...}

\item{graphs}{which plot to show ? Default is all three.}

\item{...}{...}
}
\description{
Plot modes
}
