% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_methods.R
\name{plot.Mixture}
\alias{plot.Mixture}
\title{Plot method for \code{Mixture} objects}
\usage{
\method{plot}{Mixture}(x, ...)
}
\arguments{
\item{x}{An object of class \code{Mixture}.}

\item{...}{Not used.}
}
\description{
Plot method for \code{Mixture} objects
}
