% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_methods.R
\name{plot.Mode}
\alias{plot.Mode}
\title{Plot method for \code{Mode} objects}
\usage{
\method{plot}{Mode}(modes, ...)
}
\arguments{
\item{modes}{An object of class \code{Mode}.}

\item{...}{Not used.}
}
\description{
Plot method for \code{Mode} objects
}
