% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_mcmc.R
\name{plots_mcmc}
\alias{plots_mcmc}
\title{Plots SFM MCMC output.}
\usage{
plots_mcmc(sfm_mcmc, S)
}
\arguments{
\item{sfm_mcmc}{a list. Output of `sfm_mcmc_spmix()` containing the parameter draws from the posterior distribution at each MCMC iteration.}

\item{S}{(a number between 0 and 1) The first S*M draws will be discarded as a burn-in. M is the total number of MCMC iterations.}
}
\value{
A list showing :
\itemize{
       \item 1: Trace plots of pre-processed draws;
       \item 2: The posterior distribution of each parameter after post-processing;
       \item 3: The posterior probability of the number of components after post-processing;
       \item 4: Estimated mixture density after post-processing at each iteration and histogram of the data.
}
}
\description{
Show plots of the MCMC estimation.
}
