% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_sfm_mcmc.R
\name{post_sfm_mcmc}
\alias{post_sfm_mcmc}
\title{Simple post-processing of SFM MCMC output.}
\usage{
post_sfm_mcmc(mcmc)
}
\arguments{
\item{sfm_mcmc}{a list. Output of `sfm_mcmc_spmix()` containing the parameter draws from the posterior distribution at each MCMC iteration.}

\item{burnin}{(positive integer) Number of draws used for burnin}
}
\value{
A list containing:
\itemize{
  \item A (M x 3xJb) matrix. Returns theta_draws after burn-in (discarding) the S*M rows. M is the number of rows of theta_draws (number of MCMC iterations). Jb is the number of components which are non-empty in at least one of the draws.
}
}
\description{
Gives a matrix of MCMC parameters after burn-in and discarding empty components.
}
