% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_methods.R
\name{print.BayesMode}
\alias{print.BayesMode}
\title{Print method for \code{BayesMode} objects}
\usage{
\method{print}{BayesMode}(x, max_length = 5, max_width = 5, ...)
}
\arguments{
\item{x}{An object of class \code{BayesMode}.}

\item{max_length}{maximum number of elements (for vector) or rows (for matrices) to show.}

\item{max_width}{maximum number of columns to show (for matrices).}

\item{...}{Not used.}
}
\description{
Print method for \code{BayesMode} objects
}
