% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shift_pois_mcmc.R
\name{shift_pois_mcmc}
\alias{shift_pois_mcmc}
\title{Estimation of a mixture of shifted Poisson distributions.}
\usage{
shift_pois_mcmc(y, K, nb_iter, warmup, prt = TRUE)
}
\arguments{
\item{y}{(a vector of integers) Observations used to fit the model.}

\item{K}{(an integer) Maximum number of mixture components.}

\item{nb_iter}{(an integer) Number of MCMC iterations.}

\item{prt}{print intermediate of the MCMC estimation ? default = TRUE.}
}
\value{
A list containing:
\itemize{
  \item mcmc_draws : a (nb_iter x 3xK) matrix. Parameter draws from the posterior distribution at each MCMC iteration.
  \item p_draws : a (nb_iter x K) matrix. Posterior draw of mixture weights at each MCMC iteration.
  \item kappa_draws : a (nb_iter x K) matrix. Posterior draws of kappa at each MCMC iteration.
  \item lambda_draws : a (nb_iter x K) matrix. Posterior draws of lambda at each MCMC iteration.
  \item snj : a (nb_iter x K) matrix. Number of observations allocated to each components at each MCMC iteration.
  \item se0 : a vector of size nb_iter. Concentration parameter from (symmetric) Dirichlet distribution at each MCMC iteration.
  \item dist : a string indicating the distribution used in the mixture.
  \item y, K and nb_iter, given as input.
}
}
\description{
Bayesian estimation of a mixture of shifted Poisson distributions using a Sparse Finite Mixture MCMC algorithm.
}
\references{
\insertRef{malsiner-walli_model-based_2016}{BayesMultiMode}\cr
\insertRef{basturk_bayes_2021}{BayesMultiMode}
}
\keyword{internal}
