% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_methods.R
\name{summary.bayes_mixture}
\alias{summary.bayes_mixture}
\title{Summary method for \code{bayes_mixture} objects
The summary of MCMC draws is given by the function
\code{summarise_draws} from package \pkg{posterior}.}
\usage{
\method{summary}{bayes_mixture}(object, ...)
}
\arguments{
\item{object}{An object of class \code{bayes_mixture}.}

\item{...}{Not used.}
}
\description{
Summary method for \code{bayes_mixture} objects
The summary of MCMC draws is given by the function
\code{summarise_draws} from package \pkg{posterior}.
}
