% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_methods.R
\name{summary.bayes_mode}
\alias{summary.bayes_mode}
\title{Summary method for \code{bayes_mode} objects}
\usage{
\method{summary}{bayes_mode}(object, ...)
}
\arguments{
\item{object}{An object of class \code{bayes_mode}.}

\item{...}{Not used.}
}
\description{
Summary method for \code{bayes_mode} objects
}
