% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_methods.R
\name{summary.BayesMixture}
\alias{summary.BayesMixture}
\title{Summary method for \code{BayesMixture} objects}
\usage{
\method{summary}{BayesMixture}(object, ...)
}
\arguments{
\item{object}{An object of class \code{BayesMixture}.}

\item{...}{Not used.}
}
\description{
Summary method for \code{BayesMixture} objects
}
