% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_methods.R
\name{summary.BayesMode}
\alias{summary.BayesMode}
\title{Summary modes}
\usage{
\method{summary}{BayesMode}(object, ...)
}
\arguments{
\item{object}{...}

\item{...}{...}
}
\description{
Summary modes
}
