% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_methods.R
\name{summary.Mixture}
\alias{summary.Mixture}
\title{Summary method for \code{Mixture} objects}
\usage{
\method{summary}{Mixture}(object, ...)
}
\arguments{
\item{object}{An object of class \code{Mode}.}

\item{...}{Not used.}
}
\description{
Summary method for \code{Mixture} objects
}
