% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_methods.R
\name{summary.Mode}
\alias{summary.Mode}
\title{Summary method for \code{Mode} objects}
\usage{
\method{summary}{Mode}(object, ...)
}
\arguments{
\item{object}{An object of class \code{Mode}.}

\item{...}{Not used.}
}
\description{
Summary method for \code{Mode} objects
}
