% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayes_estimation_stan.R
\name{bayes_estimation}
\alias{bayes_estimation}
\title{Bayesian estimation of mixture distributions}
\usage{
bayes_estimation(
  data,
  K,
  dist,
  nb_iter = 2000,
  burnin = nb_iter/2,
  chains = 4,
  cores = 4,
  a0 = 10,
  A0 = 10 * K,
  b0 = mean(data),
  B0 = (max(data) - min(data))^2,
  c0 = 2.5,
  e0 = 0,
  g0 = 0.5,
  G0 = 100 * 2.5/0.5/(max(data) - min(data))^2,
  h0 = 0,
  H0 = 10,
  n0 = 2,
  N0 = 0.1,
  refresh = 1000
)
}
\arguments{
\item{data}{Numeric vector of input values.}

\item{fit}{Numeric vector of output values.}

\item{...}{Arguments passed to `rstan::sampling` (e.g. iter, chains).}
}
\value{
An object of class `stanfit` returned by `rstan::sampling`
}
\description{
Bayesian estimation of mixture distributions
}
