% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayes_estimation.R
\name{bayes_estimation}
\alias{bayes_estimation}
\title{Bayesian estimation of mixture distributions}
\usage{
bayes_estimation(
  data,
  K,
  dist,
  priors = list(),
  nb_iter = 2000,
  burnin = nb_iter/2,
  printing = TRUE
)
}
\arguments{
\item{data}{Vector of observations.}

\item{K}{Maximum number of mixture components.}

\item{dist}{String indicating the distribution of the mixture components.
Currently supports "normal", "skew_normal", "poisson" and "shifted_poisson".}

\item{priors}{List of priors. Default in an empty list which implies the following priors :\cr
a0 = 1,\cr A0 = 200,\cr b0 = median(y),\cr B0 = (max(y) - min(y))^2,\cr c0 = 2.5,\cr
l0 = 1.1,\cr L0 = 1.1/median(y),\cr g0 = 0.5,\cr G0 = 100*g0/c0/B0,\cr D_xi = 1,\cr D_psi =1.}

\item{nb_iter}{Number of MCMC iterations. Default is 2000.}

\item{burnin}{Number of MCMC iterations used as burnin. Default is nb_iter/2.}

\item{printing}{Showing MCMC progression ?}
}
\value{
A list of class `BayesMixture` containing
\itemize{
 \item{data}{ - same as argument}
 \item{dist_type}{ - type of the distribution (continuous or discrete)}
 \item{pars_names}{ - names of the mixture components' parameters}
 \item{mcmc}{ - Matrix of MCMC draws where the rows corresponding to burnin have been discarded}
 \item{mcmc_all}{ - Original matrix of MCMC draws}
}
}
\description{
Gibbs sampler for Spare Finite Mixture MCMC estimation of mixture distributions.
}
\references{
\insertRef{malsiner-walli_model-based_2016}{BayesMultiMode}\cr\cr
\insertRef{fruhwirth-schnatter_bayesian_2010}{BayesMultiMode}\cr\cr
\insertRef{SFS:Mal:2019}{BayesMultiMode}\cr\cr
\insertRef{viallefont2002bayesian}{BayesMultiMode}
}
