% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayes_estimation.R
\name{bayes_estimation}
\alias{bayes_estimation}
\title{Bayesian estimation of mixture distributions}
\usage{
bayes_estimation(
  data,
  K,
  dist,
  priors = list(),
  nb_iter = 2000,
  burnin = nb_iter/2,
  printing = TRUE
)
}
\arguments{
\item{data}{Numeric vector of input values.}

\item{K}{Integer indicating the maximum number of mixture components.}

\item{dist}{String indicating the distribution of the mixture components.
Currently supports "normal", "student" and "skew_normal".}

\item{priors}{Number of MCMC iterations. Default is 2000.}

\item{nb_iter}{Number of MCMC iterations. Default is 2000.}

\item{burnin}{Number of MCMC iterations used as burnin.}

\item{printing}{Showing MCMC progression ?}
}
\value{
An object of class `BayesMixture`.
}
\description{
Bayesian estimation of mixture distributions
}
