% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayes_mode.R
\name{bayes_mode}
\alias{bayes_mode}
\title{Bayesian estimation of mixture distributions}
\usage{
bayes_mode(BayesMix, rd = 1, tol_p = 0.001, tol_x = sd(data)/10)
}
\arguments{
\item{BayesMix}{object of class `BayesMixture`.}

\item{rd}{(numeric scalar) rounding}

\item{tol_p}{...}

\item{tol_x}{...}
}
\value{
An object of class `BayesMode`.
}
\description{
Bayesian estimation of mixture distributions
}
