% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayes_mode.R
\name{bayes_mode}
\alias{bayes_mode}
\title{Bayesian estimation of mixture distributions}
\usage{
bayes_mode(BayesMix, rd = 1)
}
\arguments{
\item{BayesMix}{object of class `BayesMixture`.}
}
\value{
An object of class `BayesMode`.
}
\description{
Bayesian estimation of mixture distributions
}
