% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayes_mode.R
\name{bayes_mode}
\alias{bayes_mode}
\title{Bayesian estimation of mixture distributions}
\usage{
bayes_mode(
  BayesMix,
  rd = 1,
  tol_x = sd(BayesMix$data)/10,
  show_plot = F,
  nb_iter = NA
)
}
\arguments{
\item{BayesMix}{object of class `BayesMixture`.}

\item{rd}{Rounding parameter.}

\item{tol_x}{...}

\item{show_plot}{Show density with estimate mode as vertical bars}

\item{nb_iter}{Number of draws on which the mode-finding algorithm is run}
}
\value{
An object of class `BayesMode`.
}
\description{
Bayesian estimation of mixture distributions
}
