% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayes_mode.R
\name{bayes_mode}
\alias{bayes_mode}
\title{Bayesian estimation of mixture distributions}
\usage{
bayes_mode(
  BayesMix,
  rd = 1,
  tol_p = 0.001,
  tol_x = sd(BayesMix$data)/10,
  show_plot = F,
  nb_iter = NA
)
}
\arguments{
\item{BayesMix}{object of class `BayesMixture`.}

\item{rd}{Rounding parameter.}

\item{tol_p}{Tolerance parameter for distance in-between modes. Default is sd(data)/10. If two modes are closer than tol_x, only the first estimated mode is kept.}

\item{tol_x}{Tolerance parameter for small components. Default is 1e-3. All components with mixture weights lower than tol_p are dropped.}
}
\value{
An object of class `BayesMode`.
}
\description{
Bayesian estimation of mixture distributions
}
