% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayes_trace.R
\name{bayes_trace}
\alias{bayes_trace}
\title{Trace plots. This is wrapper around the mcmc_trace function from package bayesplot.}
\usage{
bayes_trace(BayesMix, mcmc_vars = NULL, ...)
}
\arguments{
\item{BayesMix}{An object of class BayesMixture.}

\item{mcmc_vars}{Variables to plot. Default is all the variable in the MCMC output.}

\item{...}{arguments passed to mcmc_trace.}
}
\value{
A trace plot.
}
\description{
Trace plots. This is wrapper around the mcmc_trace function from package bayesplot.
}
