% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ct47}
\alias{ct47}
\title{DNA Macrosatellite repeat ct47 (4q).
Locus (hg18): 4q35.2; Unit (kb): 3.3; Encoded product: DUX4.}
\format{
A vector of counts with 410 elements.
}
\usage{
ct47
}
\description{
DNA Macrosatellite repeat ct47 (4q).
Locus (hg18): 4q35.2; Unit (kb): 3.3; Encoded product: DUX4.
}
\references{
\insertRef{schaap_genome-wide_2013}{BayesMultiMode}
}
\keyword{datasets}
