% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cyclone}
\alias{cyclone}
\title{Tropical cylcone lifetime maximum intensity}
\format{
See https://www.ncei.noaa.gov/products/international-best-track-archive
}
\usage{
cyclone
}
\description{
Tropical cylcone lifetime maximum intensity
}
\references{
\insertRef{Richardson_Green_1997_RJMCMC}{BayesMultiMode}
}
\keyword{datasets}
