% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cyclone}
\alias{cyclone}
\title{Tropical cyclones lifetime maximum intensity}
\format{
The raw data is the ibtracs.ALL.list.v04r00.csv dataset from https://www.ncei.noaa.gov/products/international-best-track-archive
}
\source{
https://www.ncei.noaa.gov/products/international-best-track-archive
}
\usage{
cyclone
}
\description{
Tropical cyclones lifetime maximum intensity
}
\references{
\insertRef{knapp_international_2010}{BayesMultiMode}\cr\cr
\insertRef{knapp_international_2018}{BayesMultiMode}
}
\keyword{datasets}
