% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{d4z4}
\alias{d4z4}
\title{DNA data}
\format{
A vector of counts with 410 elements.
}
\usage{
d4z4
}
\description{
Macrosatellite repeat D4Z4 (4q).
Locus (hg18): 4q35.2; Unit (kb): 3.3; Encoded product: DUX4.
}
\references{
\insertRef{schaap_genome-wide_2013}{BayesMultiMode}
}
\keyword{datasets}
