% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discrete_MF.R
\name{discrete_MF}
\alias{discrete_MF}
\title{Mode-finding algorithm for mixture of discrete distributions}
\usage{
discrete_MF(mixture, type = "all")
}
\arguments{
\item{mixture}{An object of class Mixture.}

\item{type}{Type of modes, either unique or all (the latter includes flat modes); default is "all".}
}
\value{
An object of class Mode.
}
\description{
Function to estimate modes in mixtures of discrete distributions following; see Cross et al. (2023).
}
\details{
This algorithm returns the local maxima of the mixture
\deqn{p(x) = \sum_{k=1}^{K}\pi_k p_k(x).}

 By definition, modes must satisfy either: 
 \deqn{p_k(y_{m}-1) < p_k(y_{m}) > p_k(y_{m}+1)};
 \deqn{p_k(y_{m}-1) < p_k(y_{m}) = p_k(y_{m}+1) = \ldots = p_k(y_{m}+l-1) > p_k(y_{m}+l).}
 
 The algorithm evaluate each location point with these two conditions.
}
\examples{
# Example with the poisson distribution ====================================
lambda = c(0.1,10)
p = c(0.5,0.5)
params = c(eta = p, lambda = lambda)
dist = "poisson"

data = c(rpois(p[1]*1e3, lambda[1]),
         rpois(p[2]*1e3, lambda[2]))

mix = new_Mixture(params, data = data, dist = dist)

modes = discrete_MF(mix)

# summary(modes)
# plot(modes, from = 0, to = 20)

# Example with an arbitrary distribution ===================================
mu = c(20,5)
size = c(20,0.5)
p = c(0.5,0.5)
params = c(eta = p, mu = mu, size = size)

data = c(rnbinom(p[1]*1e3, mu = mu[1], size = size[1]),
         rnbinom(p[2]*1e3, mu = mu[2], size = size[2]))

pmf_func <- function(x, pars) {
  dnbinom(x, mu = pars["mu"], size = pars["size"])
}

mix = new_Mixture(params, data = data, pdf_func = pmf_func, dist_type = "discrete")
modes = discrete_MF(mix)

# summary(modes)
# plot(modes, from = 0, to = 50)

}
\references{
\insertRef{cross_2023}{BayesMultiMode}

\insertRef{schaap_genome-wide_2013}{BayesMultiMode}
}
