% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixed_point.R
\name{fixed_point}
\alias{fixed_point}
\title{Fixed-point algorithm for finding the modes of a gaussian mixture.
See Carreira-Perpinan (2000), section 4 equation (10) https://doi.org/10.1109/34.888716.}
\usage{
fixed_point(mcmc, data, tol_p = 0.001, tol_x = sd(data)/10, show_plot = F)
}
\arguments{
\item{mcmc}{A vector of estimated mixture parameters.}

\item{data}{A vector data used for estimating the mixtures.}

\item{tol_p}{Tolerance parameter for small components. Default is 1e-3. All components with mixture weights lower than tol_p are dropped.}

\item{tol_x}{Tolerance parameter for distance in-between modes. Default is sd(data)/10. If two modes are closer than tol_x, only the first estimated mode is kept.}

\item{show_plot}{Show the data and estimated modes.}
}
\value{
A vector of estimated modes.
}
\description{
Fixed-point algorithm for finding the modes of a gaussian mixture.
See Carreira-Perpinan (2000), section 4 equation (10) https://doi.org/10.1109/34.888716.
}
