% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixed_point.R
\name{fixed_point}
\alias{fixed_point}
\title{Fixed-point algorithm for finding the modes of a gaussian mixture}
\usage{
fixed_point(params, y, tol_p = 0.001, tol_x = sd(y)/10, show_plot = FALSE)
}
\arguments{
\item{params}{a vector of estimated mixture parameters.}

\item{y}{the vector of data used for estimating the mixtures.}

\item{tol_p}{Tolerance for small components. Default is 1e-3. All components with mixture weights lower than tol_p are dropped.}

\item{tol_x}{Tolerance for distance in-between modes. Default is sd(y)/10. If two modes are closer than tol_x, only the first estimated mode is kept.}

\item{show_plot}{Show the data and estimated modes.}
}
\value{
A vector of estimated modes.
}
\description{
Fixed-point algorithm to find the modes of a gaussian
mixture from Carreira-Perpinan (2000), section 4 equation (10)
https://doi.org/10.1109/34.888716
}
