% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixed_point.R
\name{fixed_point}
\alias{fixed_point}
\title{Modal fixed-point algorithm}
\usage{
fixed_point(mixture, tol_x = 1e-06, tol_conv = 1e-08)
}
\arguments{
\item{mixture}{An object of class Mixture.}

\item{tol_x}{Tolerance parameter for distance in-between modes; default is 1e-6; if two modes are closer than \code{tol_x}, only the first estimated mode is kept.}

\item{tol_conv}{Tolerance parameter for convergence of the algorithm; default is 1e-8.}
}
\value{
An object of class Mode.
}
\description{
Algorithm for estimating modes in mixture of Normal distributions from Carreira-Perpinan (2000).
}
\details{
This algorithm returns the local maxima of the mixture
\deqn{p(x) = \sum_{k=1}^{K}\pi_k p_k(x),}
where \eqn{p_k} comes from the Normal family.
Following Carreira-perpinan (2000), a mode \eqn{x} is found by iterating the two steps:
\deqn{(i) \quad p(k|x^{(n)}) = \frac{\pi_k p_k(x^{(n)})}{p(x^{(n)})},}
\deqn{(ii) \quad x^{(n+1)} = f(x^{(n)}),}
with
\deqn{f(x) = (\sum_k p(k|x) \sigma_k)^{-1}\sum_k p(k|x) \sigma_k \mu_k,}
until convergence, that is, until \eqn{abs(x^{(n+1)}-x^{(n)})< \text{tol}_\text{conv}},
where \eqn{\text{tol}_\text{conv}} is an argument with default value \eqn{1e-8}.
Following Carreira-perpinan (2000), the algorithm is started at each component location.
Separately, it is necessary to identify identical modes which diverge only up to
a small value. By default modes which are closer
than \eqn{sd(y)/10} are merged; this tolerance value can be controlled with the argument
\code{tol_x}.
}
\references{
\insertRef{carreira-perpinan_mode-finding_2000}{BayesMultiMode}
}
