% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{galaxy}
\alias{galaxy}
\title{Galaxy series}
\format{
Velocity at which 82 galaxies in the Corona Borealis region are moving away from our galaxy, scaled by 1000.
}
\usage{
galaxy
}
\description{
Galaxy series
}
\references{
\insertRef{Richardson_Green_1997_RJMCMC}{BayesMultiMode}
}
\keyword{datasets}
