% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{galaxy}
\alias{galaxy}
\title{Galaxy data set}
\format{
Univariate measurements representing velocities of galaxies moving away from our galaxy. 82 observations.
}
\usage{
galaxy
}
\description{
https://people.maths.bris.ac.uk/~mapjg/mixdat
(Richardson S, Green PJ (1997) https://doi.org/10.1111/1467-9868.00095)
}
\keyword{datasets}
