% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gibbs_sfm_algos.R
\name{gibbs_SFM_normal}
\alias{gibbs_SFM_normal}
\title{SFM MCMC algorithms to estimate mixtures.}
\usage{
gibbs_SFM_normal(y, K, nb_iter, priors = list(), print = TRUE)
}
\description{
SFM MCMC algorithms to estimate mixtures.
}
\keyword{internal}
