% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gibbs_SFM_skew_normal.R
\name{gibbs_SFM_skew_n}
\alias{gibbs_SFM_skew_n}
\title{Bayesian estimation of a mixture of Skew Normal distributions.}
\usage{
gibbs_SFM_skew_n(y, K, nb_iter, priors = list(), printing = TRUE)
}
\arguments{
\item{y}{Vector of observations.}

\item{K}{Maximum number of mixture components.}

\item{nb_iter}{Number of MCMC iterations.}

\item{priors}{List of priors. Default is :
list(a0 = 1, A0 = 200, b0 = median(y), B0 = (max(y) - min(y))^2, c0 = 2.5, g0 = 0.5, G0 = 100*g0/c0/B0, D_xi = 1, D_psi = 1)}

\item{printing}{Print intermediate of the MCMC estimation ? default = TRUE.}
}
\value{
mcmc_draws : Parameter draws from the posterior distribution at each MCMC iteration. A (nb_iter x 2K + 1) matrix.
}
\description{
MCMC estimation using a Spare Finite Mixture algorithm.
}
\references{
\insertRef{malsiner-walli_model-based_2016}{BayesMultiMode}\cr\cr
\insertRef{fruhwirth-schnatter_bayesian_2010}{BayesMultiMode}\cr\cr
\insertRef{SFS:Mal:2019}{BayesMultiMode}
}
\keyword{internal}
