% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MEM.R
\name{MEM}
\alias{MEM}
\title{Mode-finding EM algorithm (MEM)}
\usage{
MEM(
  mcmc,
  data,
  pars_names,
  dist = "NA",
  pdf_func = NULL,
  tol_x = sd(data)/10,
  show_plot = FALSE
)
}
\arguments{
\item{mcmc}{Vector of estimated mixture parameters.}

\item{data}{Vector of observations used for estimating the mixture.}

\item{pars_names}{Names of the variables mcmc draws variables.}

\item{dist}{String indicating the distribution of the mixture components. Default is "NA".
Currently supports "normal" and "skew_normal".}

\item{pdf_func}{Pdf of the mixture components associated with the mcmc draws
(if mcmc estimation has not been carried out with BayesMultiMode); default is null.}

\item{tol_x}{Tolerance parameter for distance in-between modes. Default is sd(data)/10. If two modes are closer than tol_x, only the first estimated mode is kept.}

\item{show_plot}{If true show the data and estimated modes; default is false.}
}
\value{
A vector estimated modes.

\insertRef{li_nonparametric_2007}{BayesMultiMode}
}
\description{
Mode-finding EM algorithm (MEM)
}
