% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_BayesMixture.R
\name{new_BayesMixture}
\alias{new_BayesMixture}
\title{Creating a S3 object of class `BayesMixture`.
This function is helpful for users who want to explore modes in MCMC draws which have not been
derived using the function `bayes_estimation()`.}
\usage{
new_BayesMixture(
  fit,
  data,
  dist = "NA",
  pars_name = "NA",
  density = NULL,
  dist_type = "NA"
)
}
\arguments{
\item{fit}{A matrix of MCMC draws.}

\item{data}{A vector containing the data used for estimating the model and generating the MCMC draws.}

\item{dist}{Distribution family of the mixture components supported by
the package (e.g. "normal", "student", "skew_normal", "shifted_poisson").}

\item{pars_name}{Mapping between the distribution parameters names.
This input is used only if dist_name is invalid or NULL.}

\item{density}{Pdf or pmf of the mixture components.
This input is used only if dist_name is invalid or NULL.}

\item{dist_type}{Either "continous" or "discrete"}
}
\value{
An object of class `BayesMixture`.
}
\description{
Creating a S3 object of class `BayesMixture`.
This function is helpful for users who want to explore modes in MCMC draws which have not been
derived using the function `bayes_estimation()`.
}
