% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_Mixture.R
\name{new_Mixture}
\alias{new_Mixture}
\title{Creating a S3 object of class \code{Mixture}}
\usage{
new_Mixture(
  pars,
  dist = NA_character_,
  pdf_func = NULL,
  dist_type = NA_character_,
  range = NULL,
  loc = NA_character_
)
}
\arguments{
\item{pars}{Named vector of mixture parameters.}

\item{dist}{String indicating the distribution of the mixture components; default is \code{"NA"}.
Currently supports \code{"normal"} and \code{"skew_normal"}; not needed if \code{pdf_func} is provided.}

\item{pdf_func}{Pdf of the mixture components; default is \code{NULL}.}

\item{dist_type}{Either \code{"continuous"} or \code{"discrete"}.}

\item{range}{(optional for continuous mixtures) upper and lower limit of the range where the mixture should be evaluated.}

\item{loc}{(for continuous mixtures other than Normal mixtures) String indicating the location parameter
of the distribution; the latter is used to initialise the MEM algorithm.}
}
\value{
A list of class \code{Mixture} containing:
\item{pars}{Same as argument.}
\item{dist}{Same as argument.}
\item{pdf_func}{Same as argument.}
\item{dist_type}{Same as argument.}
}
\description{
Function for creating an object of class \code{Mixture} which can subsequently be used as argument in \code{\link[=mix_mode]{mix_mode()}} to find its modes.
}
\examples{

# Example with the skew normal =============================================
xi = c(0,6)
omega = c(1,2)
alpha = c(0,0)
p = c(0.8,0.2)
params = c(eta = p, xi = xi, omega = omega, alpha = alpha)
dist = "skew_normal"

mix = new_Mixture(params, dist = dist)

# summary(mix)
# plot(mix)

# Example with an arbitrary distribution ===================================
mu = c(0,6)
omega = c(1,2)
xi = c(0,0)
nu = c(3,100)
p = c(0.8,0.2)
params = c(eta = p, mu = mu, sigma = omega, xi = xi, nu = nu)

pdf_func <- function(x, pars) {
  sn::dst(x, pars["mu"], pars["sigma"], pars["xi"], pars["nu"])
}


mix = new_Mixture(params, pdf_func = pdf_func,
dist_type = "continuous", loc = "mu")

# summary(mix)
# plot(mix, from = -4, to = 4)

}
