% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_methods.R
\name{plot.BayesMixture}
\alias{plot.BayesMixture}
\title{Plot mixture}
\usage{
\method{plot}{BayesMixture}(x, max_size = 250, transparency = 0.1, ...)
}
\arguments{
\item{x}{...}

\item{max_size}{...}

\item{transparency}{...}

\item{...}{...}
}
\description{
Plot mixture
}
