% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot.BayesMixture}
\alias{plot.BayesMixture}
\title{Plot mixture}
\usage{
\method{plot}{BayesMixture}(x, max_size = 200, tol_p = 0.001, transparency = 0.1, ...)
}
\arguments{
\item{x}{...}

\item{max_size}{...}

\item{tol_p}{...}

\item{transparency}{...}

\item{...}{...}
}
\description{
Plot mixture
}
