% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_methods.R
\name{plot.mixture}
\alias{plot.mixture}
\title{Plot method for \code{mixture} objects}
\usage{
\method{plot}{mixture}(x, from = NULL, to = NULL, ...)
}
\arguments{
\item{x}{An object of class \code{mixture}.}

\item{from}{the lower limit of the range over which the function will be plotted.}

\item{to}{the upper limit of the range over which the function will be plotted.}

\item{...}{Not used.}
}
\description{
Plot method for \code{mixture} objects
}
