% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_methods.R
\name{print.mix_mode}
\alias{print.mix_mode}
\title{Print method for \code{mix_mode} objects}
\usage{
\method{print}{mix_mode}(x, max_length = 6L, max_width = 6L, print_all = F, ...)
}
\arguments{
\item{x}{An object of class \code{mix_mode}.}

\item{max_length}{maximum number of elements (for vector) or rows (for matrices) to show. Default is \code{6L}.}

\item{max_width}{maximum number of columns to show (for matrices). Default is \code{6L}.}

\item{print_all}{override max_length and max_width to print everything? Default is FALSE.}

\item{...}{Not used.}
}
\description{
Print method for \code{mix_mode} objects
}
