% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_methods.R
\name{summary.BayesMode}
\alias{summary.BayesMode}
\title{Summary of Bayesian mode estimates.}
\usage{
\method{summary}{BayesMode}(object, ...)
}
\arguments{
\item{object}{An object of class BayesMode.}

\item{...}{Not used.}
}
\description{
Summary of Bayesian mode estimates.
}
