% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_methods.R
\name{summary.BayesMode}
\alias{summary.BayesMode}
\title{Summary method for for \code{BayesMode} objects}
\usage{
\method{summary}{BayesMode}(object, ...)
}
\arguments{
\item{object}{An object of class \code{BayesMode}.}

\item{...}{Not used.}
}
\description{
Summary method for for \code{BayesMode} objects
}
