% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_methods.R
\name{summary.mix_mode}
\alias{summary.mix_mode}
\title{Summary method for \code{mix_mode} objects}
\usage{
\method{summary}{mix_mode}(object, ...)
}
\arguments{
\item{object}{An object of class \code{mix_mode}.}

\item{...}{Not used.}
}
\description{
Summary method for \code{mix_mode} objects
}
