% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_methods.R
\name{summary.mixture}
\alias{summary.mixture}
\title{Summary method for \code{mixture} objects}
\usage{
\method{summary}{mixture}(object, ...)
}
\arguments{
\item{object}{An object of class \code{mixture}.}

\item{...}{Not used.}
}
\description{
Summary method for \code{mixture} objects
}
